requirements.txt
# python 3.10
google-generativeai
numpy
pandas
protobuf
pyperclip
scikit-learn
streamlit
watchdog
